Added Notifications
Added a way to tell that the item has
been sent, also notifications are now sent.
A notifications screen is also present.

+++ screens/Barters.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { FlatList, ScrollView } from "react-native";
+import { Alert, FlatList, ScrollView } from "react-native";
 import { db, auth } from "../firebase.conf";
 import {
     Text,
@@ -20,8 +20,8 @@ export default function Barters() {
         const barterRef = db.collection("trades");
         barterRef
             .where("initiator", "==", auth.currentUser.email)
-            .onSnapshot(snapshot => {
-                var createByUserArr = snapshot.docs.map(doc => ({
+            .onSnapshot((snapshot) => {
+                var createByUserArr = snapshot.docs.map((doc) => ({
                     ...doc.data(),
                     id: doc.id,
                 }));
@@ -29,14 +29,37 @@ export default function Barters() {
             });
         barterRef
             .where("interested", "==", auth.currentUser.email)
-            .onSnapshot(snapshot => {
-                var interestedArr = snapshot.docs.map(doc => ({
+            .onSnapshot((snapshot) => {
+                var interestedArr = snapshot.docs.map((doc) => ({
                     ...doc.data(),
                     id: doc.id,
                 }));
                 setIntinArr(interestedArr);
             });
     }, []);
+
+    const ListEmpty = () => <Subheading>This category is empty</Subheading>;
+    const SearchListEmpty = () => {
+        if (searchQuery === "") {
+            return <Subheading>Search for something...</Subheading>;
+        } else {
+            return <Subheading>No results</Subheading>;
+        }
+    };
+
+    const search = (query) => {
+        console.log(query);
+        db.collection("trades")
+            .where("name", "==", query)
+            .get()
+            .then((snapshot) => {
+                var searchRes = snapshot.docs.map(() => ({
+                    ...doc.data(),
+                    id: doc.id,
+                }));
+                setSearchResults(searchRes);
+            });
+    };
 
     /**
      * This component is rendered for all the barters the user creates.
@@ -52,23 +75,54 @@ export default function Barters() {
                 )}
             </Card.Content>
             <Card.Actions>
-                <Button
-                    onPress={() => {
-                        db.collection("trades").doc(item.id).delete();
-                    }}
-                >
-                    Cancel
-                </Button>
+                {item.fulfillment == 0 && (
+                    <Button
+                        onPress={() =>
+                            db.collection("trades").doc(item.id).delete()
+                        }
+                    >
+                        Delete
+                    </Button>
+                )}
+                {item.interested && (
+                    <Button
+                        onPressed={() => {
+                            db.collection("notifications").add({
+                                to: item.interested,
+                                from: item.initiator,
+                                timestamp: firebase.firestore.FieldValue.serverTimestamp(),
+                                msg: `${item.initiator} has sent the item.`,
+                            });
+                            if (item.fulfillment + 1 == 2) {
+                                Alert.alert(
+                                    "Both parties have sent the item, closing now."
+                                );
+                                db.collection("notifications").add({
+                                    to: item.interested,
+                                    from: item.initiator,
+                                    timestamp: firebase.firestore.FieldValue.serverTimestamp(),
+                                    msg: `Both parties have sent the item for ${item.name}`,
+                                });
+                                db.collection("trades").doc(item.id).delete();
+                            } else {
+                                db.collection("trades").doc(item.id).update({
+                                    fulfillment: firebase.firestore.FieldValue.increment(),
+                                });
+                            }
+                        }}
+                    >
+                        Sent Item
+                    </Button>
+                )}
             </Card.Actions>
         </Card>
     );
 
     /**
      * This component is rendered for all the barters the user is
-     * interested in and all others
+     * interested in
      */
-    const renderComponentForInterestedInByUserAndOthers = ({ item }) => {
-        console.log(item);
+    const renderComponentForInterestedInByUser = ({ item }) => {
         return (
             <Card>
                 <Card.Title
@@ -78,32 +132,73 @@ export default function Barters() {
                 <Card.Content>
                     <Text>{item.initiator} has started this trade</Text>
                 </Card.Content>
+                <Card.Actions>
+                    <Button
+                        onPress={() => {
+                            db.collection("notifications").add({
+                                to: item.interested,
+                                from: item.initiator,
+                                timestamp: firebase.firestore.FieldValue.serverTimestamp(),
+                                msg: `${item.initiator} has sent the item.`,
+                            });
+                            if (item.fulfillment + 1 == 2) {
+                                Alert.alert(
+                                    "Both parties have sent the item, closing now."
+                                );
+                                db.collection("notifications").add({
+                                    to: item.interested,
+                                    from: item.initiator,
+                                    timestamp: firebase.firestore.FieldValue.serverTimestamp(),
+                                    msg: `Both parties have sent the item for ${item.name}`,
+                                });
+                                db.collection("trades").doc(item.id).delete();
+                            } else {
+                                db.collection("trades").doc(item.id).update({
+                                    fulfillment: firebase.firestore.FieldValue.increment(),
+                                });
+                            }
+                        }}
+                    >
+                        Item Sent
+                    </Button>
+                </Card.Actions>
             </Card>
         );
     };
-
-    const ListEmpty = () => <Subheading>This category is empty</Subheading>;
-    const SearchListEmpty = () => {
-        if (searchQuery === "") {
-            return <Subheading>Search for something...</Subheading>;
-        } else {
-            return <Subheading>No results</Subheading>;
-        }
-    };
-
-    const search = query => {
-        console.log(query);
-        db.collection("trades")
-            .where("name", "==", query)
-            .get()
-            .then(snapshot => {
-                var searchRes = snapshot.docs.map(() => ({
-                    ...doc.data(),
-                    id: doc.id,
-                }));
-                setSearchResults(searchRes);
-            });
-    };
+    /*
+     * This component is the render component for
+     * the search results
+     */
+    function renderComponentForAll({ item }) {
+        return (
+            <Card>
+                <Card.Title
+                    title={item.name}
+                    subtitle={`For: ${item.object}`}
+                />
+                <Card.Content>
+                    <Text>{item.initiator} has started this trade</Text>
+                </Card.Content>
+                <Card.Actions>
+                    <Button
+                        onPress={() => {
+                            db.collection("trades").doc(item.id).update({
+                                interested: auth.currentUser.email,
+                            });
+                            db.collection("notifications").add({
+                                from: auth.currentUser.email,
+                                to: item.initiator,
+                                content: `${auth.currentUser.email} is willing to trade ${item.name} for ${item.name}`,
+                                on: firebase.firestore.FieldValue.serverTimestamp(),
+                            });
+                        }}
+                    >
+                        Offer to trade
+                    </Button>
+                </Card.Actions>
+            </Card>
+        );
+    }
 
     return (
         <ScrollView>
@@ -111,27 +206,27 @@ export default function Barters() {
             <FlatList
                 data={createdByUser}
                 renderItem={renderComponentForCreatedByUserArray}
-                keyExtractor={item => item.id}
+                keyExtractor={(item) => item.id}
                 ListEmptyComponent={ListEmpty}
             />
             <Title>Trades you are interested in</Title>
             <FlatList
                 data={interestedInByUser}
-                renderItem={renderComponentForInterestedInByUserAndOthers}
-                keyExtractor={item => item.id}
+                renderItem={renderComponentForInterestedInByUser}
+                keyExtractor={(item) => item.id}
                 ListEmptyComponent={ListEmpty}
             />
             <Title>All</Title>
             <Searchbar
                 placeholder="Search"
-                onChangeText={text => setSearchQuery(text)}
+                onChangeText={(text) => setSearchQuery(text)}
                 onEndEditing={search(searchQuery)}
                 value={searchQuery}
             />
             <FlatList
                 data={searchResults}
-                renderItem={renderComponentForInterestedInByUserAndOthers}
-                keyExtractor={item => item.id}
+                renderItem={renderComponentForAll}
+                keyExtractor={(item) => item.id}
                 ListEmptyComponent={SearchListEmpty}
             />
         </ScrollView>
